Refactor: move IO layer into io module (#598)
Consolidate DAC capability and playback buffer headers/sources under include/io and src/io,
and update includes/CMake accordingly.
