+++ include/daemon/audio_pipeline/audio_pipeline.h
@@ -7,8 +7,8 @@
 #include "crossfeed_engine.h"
 #include "daemon/audio_pipeline/streaming_cache_manager.h"
 #include "daemon/metrics/runtime_stats.h"
+#include "io/playback_buffer.h"
 #include "logging/logger.h"
-#include "playback_buffer.h"
 
 #include <cuda_runtime_api.h>
