+++ include/daemon/pcm/dac_manager.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "core/config_loader.h"
-#include "dac_capability.h"
+#include "io/dac_capability.h"
 
 #include <atomic>
 #include <condition_variable>
