+++ src/io/dac_capability.cpp
@@ -1,4 +1,4 @@
-#include "dac_capability.h"
+#include "io/dac_capability.h"
 
 #include <algorithm>
 #include <alsa/asoundlib.h>
