+++ tests/cpp/test_playback_buffer_threshold.cpp
@@ -1,4 +1,4 @@
-#include "playback_buffer.h"
+#include "io/playback_buffer.h"
 
 #include <gtest/gtest.h>
